common read and write code moved to protocol
